+++ .env
@@ -1,16 +1,16 @@
-REPORT_COUCH_SQS_CLIENT_CONFIG_BASIC_AUTH_USER=admin
-REPORT_COUCH_SQS_CLIENT_CONFIG_BASIC_AUTH_PASSWORD=docker
 
-REPORT_COUCH_DB_CLIENT_CONFIG_REPORT_BASIC_AUTH_USER=admin
-REPORT_COUCH_DB_CLIENT_CONFIG_REPORT_BASIC_AUTH_PASSWORD=docker
+; CouchDb Basic Auth credentials for different clients
+COUCH_DB_CLIENT_NOTIFICATION_BASIC_AUTH_USER=admin
+COUCH_DB_CLIENT_NOTIFICATION_BASIC_AUTH_PASSWORD=docker
+COUCH_DB_CLIENT_REPORT_BASIC_AUTH_USER=admin
+COUCH_DB_CLIENT_REPORT_BASIC_AUTH_PASSWORD=docker
+COUCH_DB_CLIENT_REPORT_CALCULATION_BASIC_AUTH_USER=admin
+COUCH_DB_CLIENT_REPORT_CALCULATION_BASIC_AUTH_PASSWORD=docker
 
-REPORT_COUCH_DB_CLIENT_CONFIG_REPORT_CALCULATION_BASIC_AUTH_USER=admin
-REPORT_COUCH_DB_CLIENT_CONFIG_REPORT_CALCULATION_BASIC_AUTH_PASSWORD=docker
+; SQS Basic Auth credentials
+SQS_CLIENT_BASIC_AUTH_USER=admin
+SQS_CLIENT_BASIC_AUTH_PASSWORD=docker
 
-REPORT_CHANGES_COUCH_DB_CLIENT_CONFIG_BASIC_AUTH_USER=admin
-REPORT_CHANGES_COUCH_DB_CLIENT_CONFIG_BASIC_AUTH_PASSWORD=docker
-
-NOTIFICATION_COUCH_DB_CLIENT_CONFIG_BASIC_AUTH_USER=admin
-NOTIFICATION_COUCH_DB_CLIENT_CONFIG_BASIC_AUTH_PASSWORD=docker
-
+; Encryption Key for server side secret encryption
+; e.g. webhook credentials stored encrypted in database
 CRYPTO_ENCRYPTION_SECRET=super-duper-secret
